Refactor jobmanager
