+++ apps/src/jobmanager/pkg/jobs/brokerprocessorjobcreator.py
@@ -3,9 +3,9 @@ import logging
 import uuid
 from datetime import datetime
 
-from pkg.database.database import Database
-from pkg.cache.cache import Cache
-from pkg.broker.consumer import BrokerConsumer
+from commons.database.database import Database
+from commons.cache.cache import Cache
+from commons.broker.consumer import BrokerConsumer
 from pkg.data.jobs import JobCreateRequestDto, JobDataObject
 from pkg.jobs.brokerprocessor import BrokerProcessor
