+++ apps/src/jobmanager/pkg/jobs/brokerprocessorjobscollectioncreator.py
@@ -1,9 +1,9 @@
 import json
 import logging
 
-from pkg.database.database import Database
-from pkg.cache.cache import Cache
-from pkg.broker.consumer import BrokerConsumer
+from commons.database.database import Database
+from commons.cache.cache import Cache
+from commons.broker.consumer import BrokerConsumer
 from pkg.data.jobs import OrganizationDataObject
 from pkg.jobs.brokerprocessor import BrokerProcessor
